add if exists in the migration removeCity

+++ db/migrate/20230423195911_remove_travel_city_from_general_infos.rb
@@ -1,10 +1,13 @@
 class RemoveTravelCityFromGeneralInfos < ActiveRecord::Migration[6.1]
   def change
-    remove_column :general_infos, :travel_city
-    remove_column :general_infos, :travel_state 
-    remove_column :general_infos, :travel_country
-    remove_column :general_infos, :travel_start
-    remove_column :general_infos, :travel_end
-    remove_column :general_infos, :travel_details
+    if ActiveRecord::Base.connection.column_exists?(:general_infos, :travel_city)
+      # do something if the column exists
+      remove_column :general_infos, :travel_city
+      remove_column :general_infos, :travel_state 
+      remove_column :general_infos, :travel_country
+      remove_column :general_infos, :travel_start
+      remove_column :general_infos, :travel_end
+    end
+    
   end
 end
